+++ src/com_algaworks_banco/modelo/pagamento/DocumentoPagavel.java
@@ -1,7 +1,9 @@
 package com_algaworks_banco.modelo.pagamento;
 
+import java.math.BigDecimal;
+
 public interface DocumentoPagavel {
-    double getValorTotal();
+    BigDecimal getValorTotal();
     boolean estaPago();
     void quitarPagamento();
